test: consolidate more shared test helpers and fixtures
